Subscriber list for orders
This adds the ability to specify a list of advisers to be
subscribed to an order.

The 2 endpoints are:

1. Get subscriber list:
GET /v3/omis/order/<order-id>/subscriber-list
Returns the list of advisers subscribed to an order:
[
    {
        "id": "e83a608e-84a4-11e6-ae22-56b6b6499611",
        "first_name": "Puck",
        "last_name": "Head",
        "dit_team": {
            "name": "CBBC North EAST",
            "id": "602b971d-5df6-e511-888e-e4115bead28a"
        }
    },
    {
        "id": "25628b23-c75b-4aef-b120-dac2d64c0696",
        "first_name": "",
        "last_name": "",
        "dit_team": null
    }
]

2. Change subscriber list:
PUT /v3/omis/order/<order-id>/subscriber-list
with data:
[
    {
        "id": "e83a608e-84a4-11e6-ae22-56b6b6499611"
    },
    {
        "id": "25628b23-c75b-4aef-b120-dac2d64c0696"
    }
]

Changes the list to reflect the new one removing the advisers that
are not in the list any more and adding new ones.