Add Order.service_types
This
- adds the many2many field Order.service_types
- adds service_types to the creation endpoint
- adds service_types to the update endpoint

As a service type can be disabled at any point, the endpoint
raises a validation error if the client uses a value that
was disabled at the time of creation of an order.

Disabled service types can be replaced by other service types if
they were not disabled when the order was first created.