Add manager_email and omis flag to countries
This:

- adds `manager_email` to `Country`. This is the email address
of the overseas manager overseeing that market.
- adds `omis_disabled_on` to Country. This is None if the market
offers OMIS, manages legacy records and makes sure that a market
can always be disabled in the future.
- `omis_disabled_on` is exposed via API but not `manager_email`
as the endpoint does not require authentication and that piece
of info is not really needed.
- the fixtures have been updated to include the correct value
for omis_disabled_on. All manager_emails are currently empty.