Add OMIS markets
This:
- adds a new app `market` and the related model `Market` with
country, manager_email and disabled_on fields
- adds market to the metadata registry to generate the API view
- adds the command `load_initial_omis_markets` to load the initial
markets data. As some columns are going to change via the admin,
the command only loads the records if the table is empty or if
you pass --override
- moves the OMIS fixtures to the OMIS module and adds a new command
`load_omis_metadata` to load all OMIS metadata. This is not going
to be called by `loadmetadata` so they both have to be run manually

The new apps in the `omis` folder `core` and `market` now use
a custom app name `omis-...` so that they don't collide with
potential other apps called in the same way but not part of OMIS.
This also generates tables with name `omis-...` which has nice
namespaced behaviour.
The other app `order` does not have this as it's not as straightforward.