Add OMIS notifications (#416)
This:
- adds an OMIS notifications django app
- sets up GOV.UK notifications for OMIS
- sends an email to the overseas manager of the chosen market
when a new order is created

As we don't always want the notifications to work (e.g. the
data hub team doesn't care about OMIS notifications), if
the related API_KEY is not set, the module mocks the calls.
This is also useful during tests to make sure we don't
accidentally send emails.

A couple of tests are optional and run only if a GOV.UK notification
test key is specified. They go a level deeper and check that
the templates in GOV.UK notify have not been changed.