Add ability to cancel a quote
This adds the ability to reopen an order and cancel a quote.
A new endpoint POST `v3/order/<order-id>/quote/cancel` has been
created.
The reason behind POSTing to cancel instead of DELETing the
resource is that we do want the cancelled quote to still be
referenced by the order so that it can be used to check its status.

We do not delete any quotes for the time being as they are going
to be useful in absence of audit trail.

A cancelled quote can be replaced by a new one.