Add discount value and label to OMIS orders
This adds the fields discount_value and discount_label to the
Order model.
This is just to support legacy orders and new ones will always
have zero discount for the time being.

discount_label is a descriptive text explaining the discount applied
whilst discount_value is the actual money value.