Add OMIS Order.public_token field
This adds a new field `public_token` to the order model used to
access the order via the public facing OMIS app.

The field is a random string composed of 50 uppercase, lowercase
letters, digits and symbols.
A notification with a link to the public facing OMIS app is sent
when a quote is generated.

As the public_token field is mandatory, this creates 3 migrations:

1. adds the field as optional
2. populates all the existing records
3. makes the field required