Add OMIS order get endpoint for public facing app
This adds `GET /v3/omis/public/order/<public-token>` which returns
the details of a order.
The response data is slightly different from the ordinary order
endpoints as the public facing app needs less details.
Also, the nested company is expanded as the app needs extra data
and adding an extra company endpoint is overkilled.

An order can only be accessed by its public token and only if
its status is not in draft/cancelled.