Remove expand param from quote endpoints
This removes the unnecessary logic behind the expand param to
get the content of a quote.
The frontend does not use this feature for now and it's
good to remove complexity especially as we need to add public
facing API as well.

The quote endpoints will always return its content.