Add get/accept quote via public facing API
This adds two endpoints:
GET /v3/order/public/<public_token>/quote
which returns details of a quote

POST /v3/order/public/<public_token>/quote/accept
which accepts a quote

It also changes the way an order is reopened.
Before: you could reopen an order even if it was in draft already.
After: if the order is in draft and you try to reopen it
a 409 is returned. This is to be consistent with all the other
state change methods.