Return reopened orders in public OMIS API
When an adviser reopens an order, the status changes back from
'Quote Awaiting Acceptance' to 'Draft' and the related quote cancelled.

In the meantime, the client might have received a notification asking
to accept the quote.
Instead of returning a 404 as the order is in draft, we allow the API
to return the order and the cancelled quote.
This way, the frontend can show that the order can't be progressed
as the related quote got cancelled and they are working on a new
version.