Add ability to set actual time when order is paid
Only when the order is paid, the adviser should be able to
record now many hours they actually spent on the order
(`actual_time` field).
This can be different from the estimated time they set before
starting to work on the order (`estimated_time`).

The PATCH /v3/omis/order/<order-id>/assignee endpoint is changed
to support this case.

Also, the order.completed_by and order.completed_on fields have been
added and included in the search / order endpoints.
The logic to set those will be added later on.