[DH-1210] Exclude irrelevant companies during Companies House load
This excludes primarily foreign companies and companies which are e.g. actually registered with the FCA and Companies House holds little information about.

Using the CSV dump, the company category proved to be the most reliable way of determining which companies to filter out. It's not particularly ideal as they are descriptions, but it was the best option with the CSV dump (the Companies House API on the otherhand gives more useful and granular information than the CSV dump, so this is something the API would be better for).