Add `until` filter to Investment/from endpoint
This allows narrowing the results returned by the investments/from
endpoint.

The reason for this is that datahub-mi needs to do an initial load of
investment projects from the start of the financial year. Right now
trying to request this from the endpoint hits the 30 second PaaS request
timeout. This will allow datahub-mi to pull a day/week/month at a time
until they up to date.