Change params on Investment/from endpoint
There was a discussion about the right param names for this endpoint,
and as we have the luxury of only one client using this endpoint and
that client is not live yet we can change the params without breaking
anything.

The options:

The while they sound weird the endpoint _reads_ well,
`/v3/investments/from?time=xxx&until=yyy` but this doesn't match other
places where we have date range params, e.g. search endpoints which use
`before` and `after`.

We can match the search endpoints but there is still the ambiguity at
first glance of the field you're filtering. We're making this
non-obvious at the moment. Without reading the documentation you
wouldn't know which field `before` / `after` affect.

We are not being obvious with the operator used for comparison either.
Is the API going to filter dates `>` `modified_on` or `>=` ? A consumer
can't easily guess.

We agreed to be obvious and name the field and the operator explicitly.

This commit changes the parameters to be `modified_on__lte` and `modified_on__gte`