Add ability to cancel an order via django admin
At the moment, advisers can only cancel an order if in status
draft or quote_awaiting_acceptance using the datahub-frontend.

This adds the ability for DIT staff to cancel orders in status
quote_accepted or paid as well. It belongs to the admin area
as it should be considered an exceptional case.