Add signal to sync investment projects to ES when an adviser is saved
DH-813

This is primarily to make sure the teams stored against the investment project in ES are updated.

This could potentially be a fair few records, but that shouldn't be a concern as the thread pool executor is used. We also already do something similar for contacts when a company is changed.