Add pricing fields to OMIS invoices
As we are going to allow people to change VAT fields in the future,
pricing fields that depend on them need to be stored against
invoices to keep track of their values at the time of invoice
creation.

Note that quotes only include the net cost (without VAT) so it's
safe to change the VAT status after a quote is accepted as its
net value doesn't change, only the gross one could.