Create new OMIS invoice when billing fields change
This allow advisers to edit invoice/billing details when an order
is in 'Quote Accepted' and creating a new invoice as a consequence.

The net cost (the one shown in the quote) cannot change after the
quote is accepted but the VAT details can and therefore the gross
cost on the invoice as well.