Load Companies House data without first truncating the table
DH-1471

Previously, the sync_ch command truncated the company_companieshousecompany table and then loaded the Companies House data from scratch.

This was problematic as it could only be done when the system was not in use.

This change makes the job update the table without truncating it first. This avoids the need for downtime (but with the cost of a slower update). Doing this kind of update is excessively slow when using the Django ORM, so native PostgreSQL INSERT ... ON CONFLICT SQL statements are used instead.