Avoid double-escaping in OMIS quotes
When generating the content of a quote, the logic was escaping html
chars along with markdown ones.

However Django already escapes those chars automatically when
rendering variables in templates so this caused them to be escaped
twice.

This stops escaping html chars when generating the content of a quote
to take advantage of the default Django functionality.

By default the escape_html param is True when calling the util function
escape_markdown so that it does not get misused by mistake.