Make the loadmetadata command safer
Since this shouldn't now be run in production, this renames the command to loadinitialmetadata and makes it fail if metadata records already exist (unless --force is passed).

The check for existing data was done by loading the YAML files separately and extracting a list of relevant models from it. Although not ideal, this seemed like the simplest approach. The main alternative would be to use django.core.serializers, but that goes a lot further by constructing objects from the data in the fixtures.