Add OMIS Payment Gateway Session models and logic
A Payment Gateway Session is a temporary model holding a GOV.UK
payment in progress.
This payment eventually becomes an OMIS payment.Payment if it
succeeds.
We have this separation so that we can ignore GOV.UK payments that
fail or get cancelled and because OMIS has already the notion of
a Payment object which can be a bank transfer.
This design makes it clear that a GOV.UK payment is not the
same as an OMIS payment and that a payment gateway session
is used to track it instead.

This:
- creates the Payment Gateway Session model
- creates 3 main methods:
  - create_from_order (on the session manager)
  - refresh_from_govuk_payment (on the session model)
  - cancel (on the session model)

Create from order
This method starts a GOV.UK payment cancelling other ongoing ones
if present as only one payment per order is allowed.
The order has to be in `order_accepted` state for this to happen.

Refresh from GOV.UK payment
As a GOV.UK payment is not managed by our system, we need to
call GOV.UK Pay to get the most up-to-date state.
Our Payment Gateway Session only records the status so
that we can ignore the sessions already finished (with success or
error).
When a GOV.UK payment becomes 'success', the related OMIS session
is updated, the related order is marked as 'paid' and a proper
'payment.Payment' object created from the GOV.UK payment one.

Cancel
This cancels the GOV.UK payment and the related Payment Gateway
Session if it can.