Register models with reversion via decorator
Before we were relying on the `reversion.admin.VersionAdmin`
logic which auto-registers models automatically.
This was not very safe as we are using reversion not only in
the admin but in many other places (e.g. endpoints) as well.

This removes the registration via admin and replaces it with
explicitly registering the models with reversion.

As a result, BaseModelVersionAdmin and ConfigurableVersionAdmin
are not needed any more.