Remove index from ES models
The index was being defined in a lot of places. The index defined on the model was only being used when deleting documents, so this change passes the index in delete_document() instead.

If we need to use different indexes for different models later on, we can come up with a better way of managing it.