Add sector hierarchy
DH-1585

Adds a hierarchy to sectors using django-mptt.

The migration reloads the sectors with the hierarchy in place. (Note that sectors themselves won't be modified in production while this work is ongoing.) The name column has been removed from the model, but remains in the database for now so that we can deploy without any downtime.

The main complexity here is around generating the full name/path of each sector. In theory, it's possible to do so in the same query using StringAgg (string_agg()), but getting that working with the ORM and DRF is not straightforward. So a relatively efficient approach (taking into account pre-fetching) of walking up parent attributes was used instead.

The metadata endpoint for sectors has been extended to include a few more fields. Otherwise, the API remains unchanged (i.e. these changes are backwards compatible).