Add Policy Feedback specific fields to Interaction model
To support more granular analysis of feedback this commit is adding two fields to the
Interaction Model for use only when the KIND is policy_feedback.

The Serializer supports these new fields and Serializer validation requires them
when an Interaction's type is policy_feedback, it will disallow them if the
Interaction is of any other kind.

These fields don't need indexing by ES so search.interaction.models.Interaction
will not be updated.