Return business type in Companies House view
DH-1607

This is part of the work to move the company category to business type mapping from the front-end code base to the back end.

This is so that there does not need to be a one-to-one mapping between company categories and business types, and to allow companies limited by guarantee to be added to the system.