Load communication channels fixtures in a migration
Before:
communication channel fixtures were kept in fixture files and
loaded using the loadinitialmetadata command.

However, they can be changed via the django admin so they were
really just initial data.

After:
The interaction fixtures are loaded when migrating the database
and only if the table is empty so that we don't override
existing systems.