Only store data for explicitly defined fields in Elasticsearch
DH-1658

This removes IGNORED_FIELDS from the ES models and only stores data for fields that have been explicitly defined on the model.

Fields that were previously implicitly defined (dynamically mapped by ES) have now been explicitly defined.

(Priority for investment projects was skipped as it was not added to the mapping in production, due to no data being stored in it.)