Correct company name filter for contact, interaction and order search
The company_name filter for these models was attempting to search by trading name (as well as name) but trading name was not being saved in ES for these models.

This adds trading name to those models (and also adds some tests to make sure the filter works).

This also adds a test to make sure all the composite filter fields are valid.