Move Elasticsearch index initialisation to management command
This is to stop the Elasticsearch index initialisation being done in the app ready() callback.

This is to stop the initialisation being done unnecessarily (e.g. for manage.py migrate), and also to stop the Elasticsearch connection being shared amongst Celery workers (due to how elasticsearch-dsl manages connections, and how Celery forks).