Schedule full Elasticsearch sync to run every night
DH-1660

This adds an additional Celery queue named 'long-running' for long-running tasks.

It also configures Celery to run with the -O fair --prefetch-multiplier 1 arguments for fairness when long-running tasks are pending execution.

For now the same configuration is used for all Celery worker processes, but we may decide to allocate different queues to different worker processes when needs change.