Use server-side cursors when bulk syncing to Elasticsearch
This significantly speeds up syncing very large data sets, and reduces load on the database server. The previous limit and offset method for iterating through a large query set became significantly slower as the offset got larger (placing a lot of strain on the database server).

Django does not support prefetching data when using server-side cursors, so this removes prefetches for to-many fields (and switches other fields to select-related). (For that reason, this method may be slightly slower for small data sets.)

No undesirable side effects (such as row or table locking) have been observed while a cursor is open.