Set Elasticsearch dynamic mapping mode to false
This is because the strict mode is not compatible with blue-green deployments and Elasticsearch mapping migration (e.g. a field may be being removed and a request with that field in it may hit the new index).

This changes the dynamic mapping mode to false, which still ignores extra data when saving documents, but does not make it an error.