(feat) Activity stream endpoint for pen testing
A trivial "Hello World" style endpoint, that exposes no data, but will be used
for pen testing the authentication on the activity stream endpoint.

It is required that we have 2 separate factors used for authentication. We
have:

  - Client IP checking using X-Forwarded-For header
  - Hawk authentication

In PaaS, and similar cloud providers, the client IP is in the X-Forwarded-For
header, which (apparently) has measures to prevent spoofing, and from
discussions with their support team, it is acceptable to use this in
conjunction with Hawk for authentication.

We deliberately do not use django-ipware or similar to determine the client IP.
The django-ipware library does use the X-Forwarded-For header, and I suspect it
would work, but it _also_

  - Has configured (by default) a chain of HTTP
    headers/Django request properties to fallback on if X-Forwarded-For header.
  - Has code relating to non-publically routable IP addresses
  - Has code for the X-Forwarded-For header giving it's IP addresses in a
    non-standard order.

This is all unnecessary in our case, and so has the primary effect of
increasing the attack surface area, when all that it would replace is:

   remote_address = request.META['HTTP_X_FORWARDED_FOR'].split(',')[0].strip()

Ideally we would use network-level protection on the client IP. However, this
is not possible at this time in PaaS.