Don't use nested Elasticsearch queries
The only time nested queries should be used is when:
- you have an array of objects in a field
- you need to filter on two sub-fields of that field with an AND operator (for example, company.contacts.first_name = a AND company.contacts.surname = b)

Since we have no such cases of this, and because the nested fields used in filters use include_in_parent=True, we don't need to use nested queries at all.

This replaces them with normal queries, and is one of the steps on the way to replacing most or all of our nested fields with normal objects.