Revert make links in CSV exports clickable in Excel
This is because hyperlinks to Data Hub pages don't behave correctly when clicked on in Excel. Excel tries to open the page itself and gets redirected to SSO, and then opens SSO in the browser. The user then ends up getting redirected to the Data Hub home page instead of the originally requested URL, as the front end stores the original URL in the user's session.

See https://support.microsoft.com/kb/899927 for further information.