Update to Elasticsearch 6.3
This includes updating the elasticsearch and elasticsearch-dsl libraries, and updating the docker image used for development.

There is some further clean-up that could be done around mapping type names (should now be _doc for all models) and object fields and nested documents (using the new InnerDoc class). However, those have been left as a further exercise as they are more involved changes.

The CircleCI 5.5 build has been removed and a 6.4 build added in preparation for the migration to GOV.UK PaaS-supplied Elasticsearch.