Add the ability to select duplicate companies to be merged in admin
This is the first part of a feature being added to the admin site to deal with duplicate companies.

This part allows the selection of companies that are duplicates and viewing the companies that have been selected. The selection of companies is done via the company change page so that the existing company search on the change list can be reused for this purpose.

The logic for the views has been placed in a new class named CompanyMergeViews, to help keep things organised, and also to make it easier to reuse in case we decide to use a similar pattern elsewhere.

As the feature is incomplete, it is currently behind a feature flag. The terminology used at the moment refers to merging companies, though this may change.

Subsequent parts will include the selection of which company should be kept as the active record, and the transfer of contacts and interactions from one company to another.