Add the first step for merging duplicate companies in admin
This is the first part of a feature being added to the admin site to deal with duplicate companies.

This part allows the selection of a second company to merge a duplicate company with. It includes the addition of a new form widget for this purpose.

As the feature is incomplete, it is currently behind a feature flag. The terminology used at the moment refers to merging companies, though this may change.

Subsequent parts will include the selection of which company should be kept as the active record, and the transfer of contacts and interactions from one company to another.