Update Flake8 to version 3.6.0
This includes fixing related linting errors occurring with the new version.

Note that pycodestyle has now added a rule (W504) to enforce line breaks before binary operators. This was our preferred style (and recommended in PEP 8) so it is enabled (by virtue of the rule not being in the ignore list).

Reformat assertion in investment view test