Move overseas regions fixture out from migration
When we added the OverseasRegion table, we had to create a
migration to load the data as well.

Now that all environments have this migration and therefore the
overseas region data, we can move this yaml out from the migrations
and into a fixture that we load using the loadinitialmetadata
command.