Move queuing of sync tasks for related objects to Celery worker
When saving a company, we resync all of its contacts and interactions. Similarly, saving a contact causes all of its interactions to be resynced.

As it turns out, queuing Celery tasks is relatively expensive. This moves the queuing of these tasks in the most common cases to Celery to alleviate that load from the main process (and avoid delays when responding to requests that update companies and contacts).