Return inherited Core Team in /core-team/ endpoint
This adds the methods:
- get_group_global_headquarters which returns the Global HQ
for the group that a company is part of
- get_group_core_team which returns the Core Team for the
group a company is part of

It also changes the /company/<id>/core-team endpoint to return
the Core Team of the company's Global Headquarters in case of
subsidiaries.

A Core Team is usually assigned at the Global Headquarters
level and shared among all group members.