Rename CompanyCoreTeamMember to OneListCoreTeamMember
That's because the OneList prefix should be applied in all cases
to make it more explicit and clearer that the team and its members
relate to that.

There are two migration files as Django was failing to detect that
I renamed the model with all changes in one go which would have
been problematic.