Return the duns_number field in API endpoints
This adds the read-only field duns_number to the response body
of all company endpoints.

It's currently not possible to search for a duns_number because
it's not needed and the frontend does not expose this field yet.

For similar reasons, the company export functionality does not
include the duns_number field.