Add ability to specify more filters in ModelCleanupConfig
Adds the ability to specify multiple filters, and to specify filters to use when checking for referencing
related objects. This includes splitting the delete_orphans and delete_old_records tests, as the logic
in delete_old_records will increase in complexity and it will be difficult to adapt the previous generic
tests for this.