Add OMIS order option to delete_old_records command
This command deletes orders that have not been updated in the last seven years. As there is no modified_on date for orders prior to 12 July 2014, the modification date is inferred prior to that date.

The ability to specify filters to use when checking for referencing objects was added to ModelCleanupConfig as part of this change.