Make some Company API fields read-only if duns_number set
This changes the Company PATCH API endpoint to make some of the
fields read-only if the company instance has a duns_number set.

This is to avoid users overriding values coming from an external
source.

The duns_number field can only be changed via the Django Admin
at the moment.