Add NormalizedKeyword Elasticsearch field
This is intended to be a replacement for SortableCaseInsensitiveKeywordText. The advantage is that it is a keyword field and does not use fielddata=True, and as well as the lowercase filter, it uses the asciifolding filter, so that characters with Latin counterparts (e.g. Latin letters with diacritics) are kept together.

The normaliser does not need to be manually added to the index settings as elasticsearch-dsl-py does it automatically.