Correct global search ignored models logic
Previously, /v3/search?entity=companieshousecompany was allowed but returned results for all search models. Also, despite being an ignored model, companieshousecompany was returned in the aggregations array.

This corrects the behaviour so that CompaniesHouseCompany is fully excluded from global search.

This also centralises the logic so that it is consistently applied (particularly for tests that call get_basic_search_query() directly.)