Don't use stale data when bulk syncing to Elasticsearch
This change makes bulk syncing objects to Elasticsearch use a server-side cursor only for fetching the IDs of objects that need to be synced, so that the latest data is used for objects being synced to Elasticsearch instead of a cached version of the data.

I have not observed any significant difference to performance with this change (it should alleviate some of the caching load that was put on the database before these changes, even).