Add task for replacing NULLs with the default value for a model field
This adds a Celery task that will update any NULL values for a model field with the default value for that same field.

The intention is to use this with e.g. Interaction.was_policy_feedback_provided and Interaction.policy_feedback_notes.

I was half-minded to put the task in the core app, but settled on dbmaintenance since it does fit in with the type of operations performed by other things in that app.