Make trading_name follow trading_names in search
This makes sure that the search does not use the field alias any
longer.

The trading_name is kept in the response body but its value is
now obtained from the trading_names field instead.

Also, the trading_name elasticsearch mapping is changed as it
doesn't need to be indexed any longer now that trading_names
is in use instead.