Add completed on filter to OMIS order search
There is a bit of a subtlety here: in the front end this will be a date filter (with times) but the field itself includes times. Any timestamps on the before date need to be included in the results. This is already the behaviour in Elasticsearch when you perform range queries with dates (rather than full timestamps) in the query.

I have used RelaxedDate here on the serialiser bearing in mind the above. Although it would be unlikely, if, for some reason, we decide to add time filtering as well in the future, we could potentially change the serialiser field to accept dates or timestamps.