Add name.keyword and name.trigram fields to search models
This adds `keyword` and `trigram` sub-fields to the `name` field for all models that have a `name` field.

Query logic remains unchanged so that the fields can first be populated with data. Once that is done, the query logic can be updated to use the new fields (and the old `name_keyword` and `name_trigram` fields can be removed).

These fields are all being tackled together as there is logic around `name_keyword` in `datahub.search.query_builder` that is common to all models.