Use name.keyword in search queries and remove name_keyword
This uses `name.keyword` in all search queries and removes `name_keyword` from the search models that had that field.

I am not a fan of having logic around specific fields in `datahub.search.query_builder` (not least because not all search models have a `name` field). However, for now that logic remains there.

I am also not sure why we need to use `match_phrase` with keyword fields instead of sticking with `match` throughout, but again that can be looked at separately.