Remove defunct policy feedback service and service context
This edits the migration that added the policy feedback service so that it no longer creates the service. This is to stop the service being added in new environments as it is no longer used.

The `policy_feedback` service context is also removed as it no longer serves any purpose.