Move id field search logic to SEARCH_FIELDS on search models
Previously, there was specific logic for querying the `id` field in `datahub.search.query_builder`.

This removes that logic and instead adds `id` to `SEARCH_FIELDS` on each search model. This should have the same net effect and reduces the logic tied to specific fields in `datahub.search.query_builder`.