Add many-to-many contacts field to interactions
This adds a many-to-many contacts field to the interaction model.

When the contact field is set via the API or via the admin site, the value is mirrored to the contacts field.

For now, this field is not exposed via the API or admin as the data in contact has not been fully replicated to contacts.