Add task to populate many-to-many field from foreign key
This adds a generic task that populates a many-to-many field from a foreign
key on the same model, for objects where the foreign key is not blank and
there are no existing values in the many-to-many field.

As the task is generic, it's a little bit complicated. I was hoping to
reuse it for other planned work (although this is looking unlikely now).