Popuate Interaction.contacts from Interaction.contact on start-up
This schedules the `copy_foreign_key_to_m2m_field` task to run for `Interaction.contacts` 10 minutes after app start-up.

I wasn't sure whether to put this logic in the `interaction` app or `dbmaintenance`. It seemed more logical to put it in `interaction`, but perhaps more practical to put it in `dbmaintenance`.