Use autocomplete widget for interaction contacts field in admin site
This changes the contacts field in the change form for interactions in the admin site from using a many-to-many raw ID widget to using an autocomplete field. This seems like a simple way to improve usability of the field (without resulting to complex workarounds).

Contact.__str__() now also includes the company name, as this is often useful in the admin site for disambiguation.

A related change was also made to contact admin, which now uses a query set annotation for contact names which also means that that column can be used for sorting.

We have some general problems with some admin fields not being wide enough, but I have left that as a separate task.