Add v4 of company endpoints
This adds v4 of the following endpoints:
/v4/company
/v4/company/<uuid:pk>
/v4/company/<uuid:pk>/archive
/v4/company/<uuid:pk>/audit
/v4/company/<uuid:pk>/one-list-group-core-team
/v4/company/<uuid:pk>/timeline
/v4/company/<uuid:pk>/unarchive

The main difference is that addresses are now nested and with consistent field names.

The response body of some of those endpoints did not change as they don't include
addresses but were upgraded in this instance so that the whole /v4/company/ prefix
can be migrated to v4 at once.

Also /v4/company/<uuid:pk> now only uses address (required) and registered_address
(optional) and trading address fields are deprecated.