Use registration for search views
This removes the various view attributes from search apps and instead uses decorators to add the views to a registry.

The reason for making this change is that we need to add another company search view, but the various view attributes on the search apps are starting to get a bit untidy.

This includes some related changes:

- the search app is set on search views instead of the search model (so the search model can be obtained via the app, instead of specifying it on both the view and search app)
- settings are used to partially suppress the Django search app ready logic instead of swapping out the Django apps (so that the action is more specific)
- the tests that check that various attributes on views are valid were extended to check all entity search views instead of just the main view for each app