Add InteractionDITParticipant many-to-many model
This adds a `InteractionDITParticipant` model representing a many-to-many relationship between interactions and advisers.

This is intended to replace the `dit_adviser` and `dit_team` interaction fields in due course.

When an interaction is created or updated via the API or via the admin site, the `dit_adviser` and `dit_team` fields are copied to a `InteractionDITParticipant`.

For now, the reverse relationship `dit_participants` is not exposed via the API or admin site as existing data has not been fully copied to it.