Add dit_participants to interaction search model
This adds `dit_participants` to the interaction search model.

Existing search logic continues to use `dit_adviser` and `dit_team` as we first need to release this and populate the new fields with data in the Elasticsearch database.

I took the opportunity to use InnerDoc and multi-fields. We already started migrating to multi-fields, and InnerDoc is the main suggested way of using Object in the elasticsearch-dsl docs.  (The benefits improved readability, ability to use inheritance etc.)