Remove execution of manage.py collectstatic from Procfile
We previously used `DISABLE_COLLECTSTATIC` with the build pack because the app used to attempt to connect to Elasticsearch when `collectstatic` was running, which would fail at that stage.

However, we removed the logic that caused the app to connect to Elasticsearch on app ready, so we no longer need to use `DISABLE_COLLECTSTATIC` and can remove the `collectstatic` invocation from the Procfile.

This should cut down the start-up time of each app instance slightly.